ref: add new Notification component, weather service, and flag style - data for countries exercise
