+++ Part-2/countries/src/components/CountryDetail.jsx
@@ -1,59 +1,6 @@
-import { fetchWeatherApi } from "openmeteo";
+// import { fetchWeatherApi } from "openmeteo";
 import { useEffect, useState } from "react";
-
-const url = "https://api.open-meteo.com/v1/forecast";
-
-const getWeather = async ({ params }) => {
-  const responses = await fetchWeatherApi(url, params);
-
-  const response = responses[0];
-
-  const utcOffsetSeconds = response.utcOffsetSeconds();
-
-  const current = response.current();
-
-  const weatherData = {
-    current: {
-      time: new Date((Number(current.time()) + utcOffsetSeconds) * 1000),
-      temperature2m: parseFloat(current.variables(0).value()).toFixed(2),
-      weatherCode: current.variables(1).value(),
-      windSpeed10m: parseFloat(current.variables(2).value()).toFixed(2),
-    },
-  };
-
-  return weatherData;
-};
-
-const icon = {
-  0: "https://openweathermap.org/img/wn/01d@2x.png",
-  1: "http://openweathermap.org/img/wn/02d@2x.png",
-  2: "http://openweathermap.org/img/wn/03d@2x.png",
-  3: "http://openweathermap.org/img/wn/04d@2x.png",
-  45: "http://openweathermap.org/img/wn/50d@2x.png",
-  48: "http://openweathermap.org/img/wn/50d@2x.png",
-  51: "http://openweathermap.org/img/wn/09d@2x.png",
-  53: "http://openweathermap.org/img/wn/09d@2x.png",
-  55: "http://openweathermap.org/img/wn/09d@2x.png",
-  56: "http://openweathermap.org/img/wn/09d@2x.png",
-  57: "http://openweathermap.org/img/wn/09d@2x.png",
-  61: "http://openweathermap.org/img/wn/10d@2x.png",
-  63: "http://openweathermap.org/img/wn/10d@2x.png",
-  65: "http://openweathermap.org/img/wn/10d@2x.png",
-  66: "http://openweathermap.org/img/wn/13d@2x.png",
-  67: "http://openweathermap.org/img/wn/13d@2x.png",
-  71: "http://openweathermap.org/img/wn/13d@2x.png",
-  73: "http://openweathermap.org/img/wn/13d@2x.png",
-  75: "http://openweathermap.org/img/wn/13d@2x.png",
-  77: "http://openweathermap.org/img/wn/13d@2x.png",
-  80: "http://openweathermap.org/img/wn/09d@2x.png",
-  81: "http://openweathermap.org/img/wn/09d@2x.png",
-  82: "http://openweathermap.org/img/wn/09d@2x.png",
-  85: "http://openweathermap.org/img/wn/13d@2x.png",
-  86: "http://openweathermap.org/img/wn/13d@2x.png",
-  95: "http://openweathermap.org/img/wn/11d@2x.png",
-  96: "http://openweathermap.org/img/wn/11d@2x.png",
-  99: "http://openweathermap.org/img/wn/11d@2x.png",
-};
+import weatherService from "../services/weathers";
 
 const CountryDetail = ({ country }) => {
   const [weatherData, setWeatherData] = useState(null);
@@ -68,12 +15,13 @@ const CountryDetail = ({ country }) => {
 
     console.log(params);
 
-    getWeather({ params })
+    weatherService
+      .getWeather(params)
       .then((data) => {
         setWeatherData(data);
         console.log(data);
       })
-      .catch((error) => console.log("eror weather", error));
+      .catch((error) => console.log("error weather", error));
   }, [country]);
 
   if (!weatherData) {
@@ -97,14 +45,13 @@ const CountryDetail = ({ country }) => {
           ))}
         </ul>
       </div>
-      <div>
+      <div className="flag">
         <img src={country.flags.svg} alt={country.flags.alt} />
       </div>
       <div>
         <h1>Weather in {country.name.common}</h1>
         <p>Temperature: {temperature2m}°C</p>
-        <p>Weather Code: {weatherCode}</p>
-        <img src={icon[weatherCode]} />
+        <img src={weatherCode} />
         <p>Wind Speed: {windSpeed10m} m/s</p>
       </div>
     </div>
