+++ Part-2/countries/src/components/Notification.jsx
@@ -1,0 +1,13 @@
+const Notification = ({msg}) => {
+    if(!msg) {
+        return null;
+    }
+
+    return (
+        <div>
+            {msg}
+        </div>
+    )
+}
+
+export default Notification
