+++ Part-2/countries/src/index.css
@@ -1,0 +1,6 @@
+.flag img{
+    width: 100%;
+    height: auto;
+    max-width: 10rem;
+    border: solid 0.1rem lightgray;
+}
