+++ Part-2/countries/src/main.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
+import "./index.css"
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   // <React.StrictMode>
